Homework 3,4

+++ lesson 5/lesson5.py
@@ -7,6 +7,6 @@ while True:
       print("Сумма чисел=", sum)
       break
    try:
-      sum = sum + int(text)
+      sum = sum + float(text)
    except Exception:
       print("Неправильный фориат числа, попробуйте снова")
